requirements.txt
asgi-redis
channels
daphne
Django
gunicorn==18.0
whitenoise
dj-database-url
psycopg2
djangorestframework
django-easy-rest